feat: add patch notes tab to info modal
