+++ src/locales/ja/translation.json
@@ -1,6 +1,7 @@
 {
     "about": "このゲームについて",
     "information": "情報",
+    "patchNotes": "アップデート",
     "gameName": "単語当てゲーム - {{language}}",
     "enterKey": "入力",
     "deleteKey": "削除",
