+++ src/locales/sw/translation.json
@@ -1,6 +1,7 @@
 {
   "about": "Habari ya mchezo huu",
   "information": "Taarifa",
+  "patchNotes": "Mambo Mapya",
   "gameName": "Si Word-Guessing-Game - {{language}}",
   "enterKey": "Ingiza",
   "deleteKey": "Futa",
